Fixes:    Added anti-null safe guard to prevent segfault
[git-p4: depot-paths = "//depot/qt/main/": change = 220862]

+++ src/gui/painting/qprintengine_ps.cpp
@@ -1520,6 +1520,10 @@ const ppd_file_t* QCUPSSupport::setCurrentPrinter(int index)
         _ppdClose(currPPD);
 
     const char *ppdFile = _cupsGetPPD(printers[index].name);
+
+    if (!ppdFile)
+      return 0;
+
     currPPD = _ppdOpenFile(ppdFile);
     unlink(ppdFile);
